Rename `model` folders to `models` (for consistency)
